for my friend's wedding photobooth
I have a friend who is putting together a photo booth for a wedding using a Raspberry Pi, an old printer, some used bubble gum wrappers, and a Python script courtesy yours truly. This script will continuously check the designated folder to see whether a photo has been taken (1.jpg). Once it sees a new photo, it prints out "File Found", sends the picture to the printer, and then renames and moves it to a different folder as a backup. The external print command has yet to be determined, so it's in here as a comment for now (will update later with another commit). Once the backup file is safely in its new home, it goes back to scanning for new pics. Pretty slick!