messing around with JSON to get the weather
You'll need to go to openweathermap.org and get yourself an API key (requires registration and logon) and plug it into the code here for it to work. Just run the script from your command line in bash/terminal along with whatever city name you like as an argument, and it'll get today's, tomorrow's, and the next day's weather.